Update ongoing session api to new name and refactored out ongoing session data to output folder for use in frontend
